✨ feat(common): benchmark lib
