+++ .github/workflows/rust_build.yml
@@ -22,10 +22,3 @@ jobs:
         cargo test --package heimdall -- test_ --nocapture
         cargo test --package heimdall-config -- test_ --nocapture
         cargo test --package heimdall-common -- test_ --nocapture
-
-    - name: Run Benchmarks
-      working-directory: ./heimdall
-      run: |
-        cargo test --release --package heimdall -- benchmark_ --nocapture
-        cargo test --release --package heimdall-config -- benchmark_ --nocapture
-        cargo test --release --package heimdall-common -- benchmark_ --nocapture
