+++ common/src/lib.rs
@@ -4,3 +4,4 @@ pub mod ether;
 pub mod consts;
 pub mod io;
 pub mod utils;
+pub mod testing;
